+++ Changelog.md
@@ -1,3 +1,8 @@
+## 0.0.3
+Fix issue #5 (Elevation layer tooltip localization.)
+Possible fix for issue #6 (Selecting tokens.)
+
+
 ## 0.0.2
 Fix issue #1 (Error when selecting a token in v10.284.)
 Fix issue #2 (Allow decimal steps for elevation interval.)